feat: rowHeader 圈选交互完成
